# src/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Vetsis services
host: 'localhost:8088'
basePath: /vetsis/v1
schemes:
  - http
consumes:
  - application/json
  - multipart/form-data
  - application/x-www-form-urlencoded
produces:
  - application/json
paths:
...
  /pet/deletePet:
     x-swagger-router-controller: pet
     delete:
       description: Delete a pet
       operationId: deletePet
       tags:
         - Pet
       parameters:
          - in: header
            name: pet_id
            description: Group name to delete
            type: string
            required: true
       responses:
         "200":
           description: Delete pet
         "400":
           description: Value is required but was not provide
  /pet/modifyPet:
    x-swagger-router-controller: pet
    put:
      description: Modify a pet
      operationId: modifyPet
      tags:
        - Pet
      parameters:
        - name: content
          in: body
          description: Body schema for modify a customers post
          required: true
          schema:
            $ref: '#/definitions/NewPets'
      responses:
          "200":
            description: >-
          "400":
            description: >-
  /pet/pets:
    x-swagger-router-controller: pet
    get:
      description: Return the list of available pets
      operationId: showPets
      tags:
        - Pet
      parameters:
          - in: header
            name: pet_id
            description: Group name to delete
            type: string
      responses:
        "200":
          description: Application list
          schema:
            $ref: "#/definitions/GetPets"
        "400":
          description: Value is required but was not provided
  /provider/addProvider:
     x-swagger-router-controller: provider
     post:
        description: Store new providers
        operationId: newProvider
        tags:
          - Provider
        parameters:
          - name: content
            in: body
            description: Body schema for new provider post
            required: true
            schema:
              $ref: '#/definitions/NewProviders'
        responses:
          "400":
            description: >-
  /provider/deleteProvider:
     x-swagger-router-controller: provider
     delete:
       description: Delete a provider
       operationId: deleteProvider
       tags:
        - Provider
       parameters:
          - in: header
            name: provider_id
            description: provider id to delete
            type: string
            required: true
       responses:
         "200":
           description: Delete provider
         "400":
           description: Value is required but was not provide
  /provider/providers:
    x-swagger-router-controller: provider
    get:
      description: Return the list of available providers
      operationId: showProviders
      tags:
        - Provider
      parameters:
          - in: header
            name: provider_id
            description: Group name to delete
            type: string
      responses:
        "200":
          description: Provider list
          schema:
            $ref: "#/definitions/GetProviders"
        "400":
          description: Value is required but was not provided
  /provider/orders:
    x-swagger-router-controller: provider
    get:
      description: Return the list of  orders
      operationId: showOrders
      tags:
        - Provider
      responses:
        "200":
          description: Provider list
          schema:
            $ref: "#/definitions/GetOrders"
        "400":
          description: Value is required but was not provided}
  /provider/newOrder:
    x-swagger-router-controller: provider
    post:
      description: Store new orders
      operationId: newOrder
      tags:
        - Provider
      parameters:
        - name: content
          in: body
          description: Body schema for new provider post
          required: true
          schema:
            $ref: '#/definitions/NewOrders'
      responses:
        "200":
          description: >-
        "400":
          description: >-
  /provider/deleteOrder:
     x-swagger-router-controller: provider
     delete:
       description: Delete order
       operationId: deleteOrder
       tags:
        - Provider
       parameters:
          - in: header
            name: order_id
            description: order id to delete
            type: string
            required: true
       responses:
         "200":
           description: Delete order
         "400":
           description: Value is required but was not provide
...
  /breed/addBreed:
      x-swagger-router-controller: breed
      post:
        description: Store new breed
        operationId: newBreed
        tags:
          - Breed
        parameters:
          - name: content
            in: body
            description: Body schema for new breed post
            required: true
            schema:
              $ref: '#/definitions/NewBreed'
        responses:
          "400":
            description: >-
  /breed/deleteBreed:
      x-swagger-router-controller: breed
      delete:
        description: Delete a breed
        operationId: deleteBreed
        tags:
          - Breed
        parameters:
          - in: header
            name: breed_id
            description: Breed id to delete
            type: string
            required: true
        responses:
          "200":
            description: Delete breed
          "400":
            description: Value is required but was not provided
  /breed/modifyBreed/:
    x-swagger-router-controller: breed
    put:
      description: Modify a breed
      operationId: modifyBreed
      tags:
        - Breed
      parameters:
        - name: content
          in: body
          description: Body schema for modify a breed post
          required: true
          schema:
            $ref: '#/definitions/NewBreed'
      responses:
        "400":
          description: >-
  /breed/breeds:
      x-swagger-router-controller: breed
      get:
        description: Return the list of available breeds
        operationId: showBreed
        tags:
          - Breed
        parameters:
          - in: header
            name: species_id
            description: Breed id to show
            type: string
        responses:
          "200":
            description: Application list
            schema:
              $ref: "#/definitions/GetBreeds"
          "400":
            description: Value is required but was not provided
  /category/addCategory:
      x-swagger-router-controller: category
      post:
        description: Store new category
        operationId: newCategory
        tags:
          - Category
        parameters:
          - name: content
            in: body
            description: Body schema for new category post
            required: true
            schema:
              $ref: '#/definitions/NewCategory'
        responses:
          "200":
            description: >-
          "400":
            description: >-
  /category/deleteCategory:
      x-swagger-router-controller: category
      delete:
        description: Delete a category
        operationId: deleteCategory
        tags:
          - Category
        parameters:
          - in: header
            name: category_id
            description: Category id to delete
            type: string
            required: true
        responses:
          "204":
            description: Delete category
          "400":
            description: Value is required but was not provided
  /category/modifyCategory:
    x-swagger-router-controller: category
    put:
      description: Modify a category
      operationId: modifyCategory
      tags:
        - Category
      parameters:
        - name: content
          in: body
          description: Body schema for modify a category post
          required: true
          schema:
            $ref: '#/definitions/NewCategory'
      responses:
        "200":
          description: >-
        "400":
          description: >-
  /category/categories:
      x-swagger-router-controller: category
      get:
        description: Return the list of available categories
        operationId: showCategories
        tags:
          - Category
        parameters:
          - in: header
            name: category_id
            type: string
        responses:
          "200":
            description: Application list
            schema:
              $ref: "#/definitions/GetCategories"
          "400":
            description: Value is required but was not provided
  /product/addProduct:
      x-swagger-router-controller: product
      post:
        description: Store new product
        operationId: newProduct
        tags:
          - Products
        parameters:
          - name: content
            in: body
            description: Body schema for new product post
            required: true
            schema:
              $ref: '#/definitions/NewProducts'
        responses:
          "200":
            description: >-
          "400":
            description: >-
  /product/deleteProduct:
      x-swagger-router-controller: product
      delete:
        description: Delete a product
        operationId: deleteProduct
        tags:
          - Products
        parameters:
          - in: header
            name: product_id
            description: Product id to delete
            type: string
            required: true
        responses:
          "204":
            description: Delete product
          "400":
            description: Value is required but was not provided
  /product/modifyProduct:
    x-swagger-router-controller: product
    put:
      description: Modify a product
      operationId: modifyProduct
      tags:
        - Products
      parameters:
        - name: content
          in: body
          description: Body schema for modify a product post
          required: true
          schema:
            $ref: '#/definitions/NewProducts'
      responses:
        "200":
            description: >-
        "400":
          description: >-
  /product/products:
      x-swagger-router-controller: product
      get:
        description: Return the list of available products
        operationId: showProducts
        tags:
          - Products
        parameters:
          - in: header
            name: product_id
            description: Product id to show
            type: string
        responses:
          "200":
            description: Application list
            schema:
              $ref: "#/definitions/GetProducts"
          "400":
            description: Value is required but was not provided
  /species/addSpecies:
      x-swagger-router-controller: species
      post:
        description: Store new species
        operationId: newSpecies
        tags:
          - Species
        parameters:
          - name: content
            in: body
            description: Body schema for new species post
            required: true
            schema:
              $ref: '#/definitions/NewSpecies'
        responses:
          "400":
            description: >-
  /species/deleteSpecies:
      x-swagger-router-controller: species
      delete:
        description: Delete a species
        operationId: deleteSpecies
        tags:
          - Species
        parameters:
          - in: header
            name: species_id
            description: Species id to delete
            type: string
            required: true
        responses:
          "200":
            description: Delete species
          "400":
            description: Value is required but was not provided
  /species/modifySpecies/:
    x-swagger-router-controller: species
    put:
      description: Modify a species
      operationId: modifySpecies
      tags:
        - Species
      parameters:
        - name: content
          in: body
          description: Body schema for modify a species post
          required: true
          schema:
            $ref: '#/definitions/NewSpecies'
      responses:
        "400":
          description: >-
  /species/species:
      x-swagger-router-controller: species
      get:
        description: Return the list of available species
        operationId: showSpecies
        tags:
          - Species
        parameters:
          - in: header
            name: species_id
            description: Species id to show
            type: string
        responses:
          "200":
            description: Application list
            schema:
              $ref: "#/definitions/GetSpecies"
          "400":
            description: Value is required but was not provided
...
definitions:
  GetSpecies:
    description: GetSpecies response
    type: string
  GetProducts:
    description: GetProducts response
    type: string
  GetCategories:
    description: GetCategories response
    type: string
...
  GetBreeds:
    description: GetBreeds response
    type: string
  GetOrders:
    description: GetOrders response
  GetProviders:
    description: GetProviders response
    type: string
...
  GetPets:
    description: GetPets response
    type: string
...
  NewProducts:
    description: Body schema for new product post
    properties:
      id:
        $ref: '#/definitions/customerIdBodyParam'
      name:
        $ref: '#/definitions/vaccinationsNameBodyParam'
      idCategory:
        $ref: '#/definitions/customerIdBodyParam'
      quantity:
        $ref: '#/definitions/customerIdBodyParam'
      price_one:
        $ref: '#/definitions/customerIdBodyParam'
      price_two:
        $ref: '#/definitions/customerIdBodyParam'
      price_three:
        $ref: '#/definitions/customerIdBodyParam'
      price_four:
        $ref: '#/definitions/customerIdBodyParam'
      minimum_quantity:
        $ref: '#/definitions/customerIdBodyParam'
    required:
     - name
     - idCategory
     - quantity
     - price_one
     - minimum_quantity
  NewSpecies:
    description: Body schema for new species post
    properties:
      id:
        $ref: '#/definitions/customerIdBodyParam'
      name:
        $ref: '#/definitions/vaccinationsNameBodyParam'
    required:
     - name
  NewBreed:
    description: Body schema for new breed post
    properties:
      id:
       $ref: '#/definitions/customerIdBodyParam'
      idSpecies:
       $ref: '#/definitions/customerIdBodyParam'
      name:
        $ref: '#/definitions/vaccinationsNameBodyParam'
    required:
     - idSpecies
     - name
...
  NewProviders:
    description: Body schema for new provider post
    properties:
      id:
        $ref: '#/definitions/providerIdBodyParam'
      name:
        $ref: '#/definitions/providerNameBodyParam'
      address:
        $ref: '#/definitions/providerAddressBodyParam'
      cellPhone:
        $ref: '#/definitions/providerCellPhoneBodyParam'
    required:
      - name
      - address
      - cellPhone
  NewOrders:
    description: Body schema for new order post
    properties:
      idProvider:
        $ref: '#/definitions/providerIdBodyParam'
      amount:
        $ref: '#/definitions/orderAmountBodyParam'
      observations:
        $ref: '#/definitions/providerCellPhoneBodyParam'
    required:
      - idProvider
      - amount
...
  NewPets:
    description: Body schema for new pet post
    properties:
      id:
        $ref: '#/definitions/customerIdBodyParam'
      idCustomer:
        $ref: '#/definitions/customerIdBodyParam'
      name:
        $ref: '#/definitions/petNameBodyParam'
      kind:
        $ref: '#/definitions/petRaceBodyParam'
      race:
        $ref: '#/definitions/petRaceBodyParam'
      gender:
        $ref: '#/definitions/petGenderBodyParam'
      castrated:
        $ref: '#/definitions/petCastratedBodyParam'
      birthdayDate:
        $ref: '#/definitions/petBirthdayDateBodyParam'
      idSpecies:
        $ref: '#/definitions/customerIdBodyParam'
      idBreed:
        $ref: '#/definitions/customerIdBodyParam'
    required:
      - name
      - idCustomer
      - kind
      - race
      - gender
      - birthdayDate
...
  NewCategory:
    description: Body schema for new customer post
    properties:
      name:
        $ref: '#/definitions/customerNameBodyParam'
...
  customerIdBodyParam:
    description: Customer id
    type: integer
    minLength: 1
    example: 1
  customerNameBodyParam:
    description: Customers name
    type: string
    minLength: 1
    example: Nicolas
...
  petNameBodyParam:
    description: Pet name
    type: string
    minLength: 1
    example: Laika
  petRaceBodyParam:
    description: Customers Location
    type: integer
    minLength: 1
    example: 0
  petGenderBodyParam:
    description: Customers Location
    type: integer
    minLength: 1
    example: 0
  petCastratedBodyParam:
    type: boolean
    minLength: 1
    example: false
  petBirthdayDateBodyParam:
    type: string
    minLength: 1
    example: "17/09/1989"
  providerIdBodyParam:
    type: string
    minLength: 1
    example: "1"
  providerNameBodyParam:
    type: string
    minLength: 1
    example: "Purina"
  providerAddressBodyParam:
    type: string
    minLength: 1
    example: "Catamarca 1880"
  providerCellPhoneBodyParam:
    type: string
    minLength: 1
    example: "3516536411"
...
  vaccinationsNameBodyParam:
    type: string
    minLength: 1
    example: "Rabia"
...
  orderAmountBodyParam:
    type: string
    minLength: 1
    example: "45.1"